+++ stories/primer/underline_nav_component_stories.rb
@@ -12,15 +12,15 @@ class Primer::UnderlineNavComponentStories < ViewComponent::Storybook::Stories
     end
 
     content do |c|
-      c.tab(selected: true) do |t|
+      c.tab(selected: true, id: "tab-1") do |t|
         t.panel { "Panel 1" }
         t.text { "Tab 1" }
       end
-      c.tab do |t|
+      c.tab(id: "tab-2") do |t|
         t.panel { "Panel 2" }
         t.text { "Tab 2" }
       end
-      c.tab do |t|
+      c.tab(id: "tab-3") do |t|
         t.panel { "Panel 3" }
         t.text { "Tab 3" }
       end