+++ CHANGELOG.md
@@ -44,6 +44,10 @@ The category for changes related to documentation, testing and tooling. Also, fo
 
 ### Breaking changes
 
+* Ensure panels in `Navigation::Tab` have a label.
+
+    *Kate Higa*
+
 * Restrict tag for `Popover` to `:div` and `Popover` heading slot to headings.
 
     *Kate Higa*